requirements.txt
flask==3.0.0
flask-cors==4.0.0
opencv-python==4.8.1.78
numpy==1.26.2
psycopg2-binary
python-dotenv==1.0.0
requests==2.31.0
Pillow
sqlalchemy==2.0.35
qrcode[pil]==7.4.2
